requirements.txt
streamlit==1.36.0
numpy==1.26.2
pandas==2.1.4
Pillow==10.1.0